requirements.txt
python-binance==1.0.17
pandas==2.0.3
numpy==1.24.3
PyQt5==5.15.9
python-dotenv==1.0.0
requests==2.31.0
ta==0.10.2
websocket-client==1.8.0
pytest==8.2.2
pytest-qt==4.4.0
ruff==0.5.7
prometheus_client==0.20.0
jsonschema==4.19.2

# ML dependencies
scikit-learn==1.4.2
scipy==1.11.4
